scraper/requirements.txt
celery==5.3.6
redis==5.0.1
sqlalchemy>=2.0.30
pg8000==1.30.5
httpx==0.27.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
